+++ src/components/BulkEditPane/BulkEditListResult/BulkEditInApp/ContentUpdatesForm/helpers.test.js
@@ -668,7 +668,7 @@ describe('ContentUpdatesForm helpers', () => {
           option: OPTIONS.HOLDINGS_NOTE,
           options: [],
           formatMessage,
-          capability: CAPABILITIES.HOLDING
+          capability: CAPABILITIES.HOLDING,
         })))
           .toEqual(
             JSON.stringify({
@@ -719,6 +719,7 @@ describe('ContentUpdatesForm helpers', () => {
                   },
                   [ACTION_VALUE_KEY]: '',
                   [FIELD_VALUE_KEY]: '',
+                  [ACTION_PARAMETERS_KEY]: []
                 },
               ],
             }),
@@ -943,6 +944,66 @@ describe('ContentUpdatesForm helpers', () => {
           expect(JSON.stringify(result))
             .toEqual(JSON.stringify(expectedStructure));
         });
+      });
+
+      it('returns the correct object for the ITEM_NOTE for ITEM option', () => {
+        expect(JSON.stringify(getDefaultActions({
+          option: OPTIONS.ITEM_NOTE,
+          formatMessage: mockFormattedMessage,
+          options: [],
+          capability: CAPABILITIES.ITEM
+        })))
+          .toEqual(
+            JSON.stringify({
+              type: '',
+              actions: [
+                null,
+                {
+                  actionsList: [{
+                    value: '',
+                    disabled: true,
+                    label: undefined
+                  }, {
+                    value: ACTIONS.MARK_AS_STAFF_ONLY,
+                    disabled: false,
+                    label: undefined
+                  }, {
+                    value: ACTIONS.REMOVE_MARK_AS_STAFF_ONLY,
+                    disabled: false,
+                    label: undefined
+                  }, {
+                    value: ACTIONS.ADD_TO_EXISTING,
+                    disabled: false,
+                    label: undefined
+                  },
+                  {
+                    value: ACTIONS.REMOVE_ALL,
+                    disabled: false,
+                    label: undefined
+                  },
+                  {
+                    value: ACTIONS.FIND,
+                    disabled: false,
+                    label: undefined
+                  },
+                  {
+                    value: ACTIONS.CHANGE_TYPE,
+                    disabled: false,
+                    label: undefined,
+                  }],
+                  [ACTION_VALUE_KEY]: '',
+                  [FIELD_VALUE_KEY]: '',
+                  [ACTION_PARAMETERS_KEY]: [
+                    {
+                      key: PARAMETERS_KEYS.STAFF_ONLY,
+                      value: false,
+                      onlyForActions: [ACTIONS.ADD_TO_EXISTING]
+                    },
+                  ]
+                },
+              ],
+            }),
+          );
       });
 
       it('should return an empty array for non-matching OPTIONS and ACTIONS', () => {